refactor: Update imports for `logging.formatter.pretty()`
